:truck: Move ObserverNavigationNode directly to cosmoscout directory
